GRP-6952 - Fix NPE when reports don't have file pointers

+++ grouper/src/grouper/edu/internet2/middleware/grouper/app/reports/GrouperReportLogic.java
@@ -516,9 +516,11 @@ public class GrouperReportLogic {
   public static void deleteFromFileSystem(GrouperReportInstance reportInstance) {
     
     String filePath = reportInstance.getReportInstanceFilePointer();
-    File file = new File(filePath);
-    if (file.exists()) {
-      file.delete();
+    if (StringUtils.isNoneBlank(filePath)) {
+      File file = new File(filePath);
+      if (file.exists()) {
+        file.delete();
+      }
     }
   }
